MA-7987 setup a unittest which trigger low memory killer
1.Setup a testCase, which include serveral testMethod
These test include test with run a browser, camera and
gallery for a loop. It will last for almost 2.5-3 hours.It will
alloc buffer via gralloc when these apps run and try to launch
a test activity.If the system hang, it will FAIL casued by timeout.
If the system or framework reboot, it will FAIL caused by process
crashed.If the testActivity can not be started normally,it will FAIL.

2.Setup a AllocBufferService which can alloc buffer via gralloc.
The AllocBufferTestCase will start this service when the test
runs and stop it after the test is over.

To run this test, some preconditions should be done:
1.adb install the CtsIMxTestApp.apk
2.copy the cts-media clips to board.
And then run cts -c com.imx.cts.AllocBufferTest or run cts -p android.imxCts

Change-Id: I4994c0b4cfe81f216b3edd58f6db7260c233b2da
Signed-off-by: Jiamin Ma<b55507@freescale.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/193
Reviewed-by: Xiaowen Liu <xiaowen.liu@nxp.com>